PERMISOS
MODIFICANDO PERMISOS, PENDIENTE HACER LA CONFIGURACION EN GASTOS CONTROLLER.

+++ app/Http/Controllers/DetalleDevolucionesCompraController.php
@@ -29,139 +29,163 @@ class DetalleDevolucionesCompraController extends Controller
 	
 	public function store(Request $request)
 	{
-	if($request->session()->has('id'))
-	 {
-	   
-  if($request->session()->get('perfil')==1  )
-			   {
-
-if($request->get('tipoproducto')==2)
+	
+	
+		if($request->session()->has('id'))
 		{
-			
-			$subtotalcant=$request->get('preciosugerido')/$request->get('cantidadempaque')*$request->get('cant');
-	        $subtotaluni=$request->get('und')*$request->get('preciosugerido');
-			$subtotal=$subtotalcant+$subtotaluni;
-			echo "Subtotal".$subtotal; echo "<br>";
-			
-		   $cantidad=$request->get('cantidadempaque')*$request->get('und')+$request->get('cant');
-		  
-		   echo "Cantidad".$cantidad; echo "<br>";
-		   
-		   
-		   $valorventa=$subtotal*$request->get('iva')/100+$subtotal;
-		   echo "Valor venta".$valorventa; echo "<br>";
-		   
-		   
-		    $iddevolucion=$request->get('iddevolucion');
-			echo "Idventa".$iddevolucion; echo "<br>";
-			
-			 $idproducto=$request->get('idproducto');
-			echo "Idproducto".$idproducto; echo "<br>";
-			
-			$utilidadcant=$request->get('utilidad')/$request->get('cantidadempaque')*$request->get('cant');
-			$combiva=$valorventa;
-			$utilidaduni=$request->get('und')*$request->get('utilidad');
-			$sumacantidad=$utilidadcant+$utilidaduni;
-			$utilidadtotal=$subtotal-$sumacantidad;
-			echo $utilidadtotal; 
-			$codigo=$request->get('codigo');
-			echo "Utilidades". $utilidadtotal; 
-			
-		   }
-		   else
+   
+		   $permiso=DB::table('permiso as p')
+			   ->where('p.idrol','=',$request->session()->get('perfil'))
+			   ->where('p.idrecurso','=',9)
+			   ->first();
+	   if(count($permiso)==0)
+	   {
+		   return Redirect::to('peticion/error')->with('mensaje','No existe ningun recurso disponible para este empleado');
+	   }else
+   
+	   {
+   
+   
+		   if($permiso->crear==1  )
 		   {
-			$subtotaluni=$request->get('preciosugerido')*$request->get('und'); 
-			$subtotal=$subtotaluni;
-			echo "Subtotal".$subtotal; echo "<br>"; 
-			
-		   $cantidad=$request->get('und');
-		   
-		   echo "Cantidad".$cantidad; echo "<br>"; 
-		   
-		   
-		   $valorventa=$subtotal*$request->get('iva')/100+$subtotal;
-		   echo "Valor venta".$valorventa; echo "<br>"; 
-		   
-		   
-		    $iddevolucion=$request->get('iddevolucion');
-			echo "Idventa".$iddevolucion; echo "<br>";
-			
-			 $idproducto=$request->get('idproducto');
-			echo "Idproducto".$idproducto; echo "<br>";
-			
-		
-			$combiva=$valorventa;
-			$utlidaduni=$request->get('und')*$request->get('utilidad');
-			$utilidadtotal=$utlidaduni-$subtotal;
-			$codigo=$request->get('codigo');
-
-			
-			echo "Utilidades".$utilidadtotal; 
-		
-			   
-		   }
-		 
-		   $detalledevolucion=DB::table('detalledevolucioncompra as d')
-						  ->join('producto as p','d.idproducto','=','p.idproducto')
-						  ->join('proveedor as pr','pr.idproveedor','=','p.idproveedor')
-						  ->select('p.descripcionproducto','p.codigobarra1','pr.nombreproveedor','d.cantidad','d.valor','d.subtotal','p.idproducto','d.iddetalledevolucioncompra')
-		                  ->where('d.idcompra','=',$iddevolucion)
-						  ->where('p.idproducto','=',$idproducto)
-						  ->first();
-			if(sizeof($detalledevolucion)<>0)
-			{
-				 return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto)->with('mensaje','Producto ya se encuentra en el detalle.');
-					
-			}
-			else{
-		   if($request->get('cantidad_venta')<$cantidad)
-				{
-					 return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto)->with('mensaje','La cantidad es mayor a la compra');
-					
-				}
-				else
-				{
-		   
-		   if($cantidad<=0)
-		   {
-			   
-			   return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto)->with('mensaje','La cantidad no puede ser inferior a cero');
-		   }
-		   else{
-			    
-	   $detalledevoluciones=new DetalleDevolucionesCompras;
-	   $detalledevoluciones->valor=$valorventa;
-	   $detalledevoluciones->idproducto=$idproducto;
-	   $detalledevoluciones->cantidad=$cantidad;
-	   $detalledevoluciones->subtotal=$subtotal;
-	   $detalledevoluciones->idcompra=$iddevolucion;
-	   $detalledevoluciones->utilidad=$utilidadtotal;
-	   $detalledevoluciones->save();
-	   return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto);
-			   
-			   
-			   }
-			   
-		   
-
-		   }
-		   
-		   }
-
-                                     }   
-			   Else
+   
+   
+if($request->get('tipoproducto')==2)
 {
- return Redirect::to('peticion/error')->with('mensaje','No tiene permisos necesarios para acceder a este contenido, ingresa como administrador');
-				   
-			   }
-		   
-		   
-		   }
-		   else
-		   {
- return Redirect::to('peticion/login')->with('mensaje','Debes ingresar tu cuenta para acceder');
-		   }
+	
+	$subtotalcant=$request->get('preciosugerido')/$request->get('cantidadempaque')*$request->get('cant');
+	$subtotaluni=$request->get('und')*$request->get('preciosugerido');
+	$subtotal=$subtotalcant+$subtotaluni;
+	echo "Subtotal".$subtotal; echo "<br>";
+	
+   $cantidad=$request->get('cantidadempaque')*$request->get('und')+$request->get('cant');
+  
+   echo "Cantidad".$cantidad; echo "<br>";
+   
+   
+   $valorventa=$subtotal*$request->get('iva')/100+$subtotal;
+   echo "Valor venta".$valorventa; echo "<br>";
+   
+   
+	$iddevolucion=$request->get('iddevolucion');
+	echo "Idventa".$iddevolucion; echo "<br>";
+	
+	 $idproducto=$request->get('idproducto');
+	echo "Idproducto".$idproducto; echo "<br>";
+	
+	$utilidadcant=$request->get('utilidad')/$request->get('cantidadempaque')*$request->get('cant');
+	$combiva=$valorventa;
+	$utilidaduni=$request->get('und')*$request->get('utilidad');
+	$sumacantidad=$utilidadcant+$utilidaduni;
+	$utilidadtotal=$subtotal-$sumacantidad;
+	echo $utilidadtotal; 
+	$codigo=$request->get('codigo');
+	echo "Utilidades". $utilidadtotal; 
+	
+   }
+   else
+   {
+	$subtotaluni=$request->get('preciosugerido')*$request->get('und'); 
+	$subtotal=$subtotaluni;
+	echo "Subtotal".$subtotal; echo "<br>"; 
+	
+   $cantidad=$request->get('und');
+   
+   echo "Cantidad".$cantidad; echo "<br>"; 
+   
+   
+   $valorventa=$subtotal*$request->get('iva')/100+$subtotal;
+   echo "Valor venta".$valorventa; echo "<br>"; 
+   
+   
+	$iddevolucion=$request->get('iddevolucion');
+	echo "Idventa".$iddevolucion; echo "<br>";
+	
+	 $idproducto=$request->get('idproducto');
+	echo "Idproducto".$idproducto; echo "<br>";
+	
 
+	$combiva=$valorventa;
+	$utlidaduni=$request->get('und')*$request->get('utilidad');
+	$utilidadtotal=$utlidaduni-$subtotal;
+	$codigo=$request->get('codigo');
+
+	
+	echo "Utilidades".$utilidadtotal; 
+
+	   
+   }
+ 
+   $detalledevolucion=DB::table('detalledevolucioncompra as d')
+				  ->join('producto as p','d.idproducto','=','p.idproducto')
+				  ->join('proveedor as pr','pr.idproveedor','=','p.idproveedor')
+				  ->select('p.descripcionproducto','p.codigobarra1','pr.nombreproveedor','d.cantidad','d.valor','d.subtotal','p.idproducto','d.iddetalledevolucioncompra')
+				  ->where('d.idcompra','=',$iddevolucion)
+				  ->where('p.idproducto','=',$idproducto)
+				  ->first();
+	if(sizeof($detalledevolucion)<>0)
+	{
+		 return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto)->with('mensaje','Producto ya se encuentra en el detalle.');
+			
+	}
+	else{
+   if($request->get('cantidad_venta')<$cantidad)
+		{
+			 return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto)->with('mensaje','La cantidad es mayor a la compra');
+			
+		}
+		else
+		{
+   
+   if($cantidad<=0)
+   {
+	   
+	   return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto)->with('mensaje','La cantidad no puede ser inferior a cero');
+   }
+   else{
+		
+$detalledevoluciones=new DetalleDevolucionesCompras;
+$detalledevoluciones->valor=$valorventa;
+$detalledevoluciones->idproducto=$idproducto;
+$detalledevoluciones->cantidad=$cantidad;
+$detalledevoluciones->subtotal=$subtotal;
+$detalledevoluciones->idcompra=$iddevolucion;
+$detalledevoluciones->utilidad=$utilidadtotal;
+$detalledevoluciones->save();
+return Redirect::to('peticion/devolucionescompras/'.$iddevolucion.'?p='.$idproducto);
+	   
+	   
+	   }
+	   
+   
+
+   }
+   
+   }
+   
+								 }   
+		   else
+   {
+   return Redirect::to('peticion/error')->with('mensaje','No tiene permisos necesarios para acceder a este contenido, ingresa como administrador');
+			   
+		   }
+   
+   
+   
+   
+		   
+	   }
+   
+		  
+	 
+			  
+			  
+			  }
+			  else
+			  {
+	return Redirect::to('peticion/login')->with('mensaje','Debes ingresar tu cuenta para acceder');
+			  }
+   
 	  
   
 
@@ -177,38 +201,55 @@ if($request->get('tipoproducto')==2)
 	}
 	 public function destroy(Request $request,$id)
    {
-	  if($request->session()->has('id'))
-	 {
-	   
-  if($request->session()->get('perfil')==1  )
-			   {
+	 
+	
+	if($request->session()->has('id'))
+	{
 
-$detalle=DetalleDevolucionesCompras::findOrFail($id);
-	 $detalle->delete();
-	 return Redirect::to('peticion/devolucionescompras/'.$detalle->idcompra);
+	   $permiso=DB::table('permiso as p')
+		   ->where('p.idrol','=',$request->session()->get('perfil'))
+		   ->where('p.idrecurso','=',9)
+		   ->first();
+   if(count($permiso)==0)
+   {
+	   return Redirect::to('peticion/error')->with('mensaje','No existe ningun recurso disponible para este empleado');
+   }else
 
-                                     }   
-			   Else
+   {
+
+
+	   if($permiso->crear==1 )
+	   {
+
+
+		$detalle=DetalleDevolucionesCompras::findOrFail($id);
+		$detalle->delete();
+		return Redirect::to('peticion/devolucionescompras/'.$detalle->idcompra);
+
+							 }   
+	   else
 {
- return Redirect::to('peticion/error')->with('mensaje','No tiene permisos necesarios para acceder a este contenido, ingresa como administrador');
-				   
-			   }
+return Redirect::to('peticion/error')->with('mensaje','No tiene permisos necesarios para acceder a este contenido, ingresa como administrador');
 		   
-		   
-		   }
-		   else
-		   {
- return Redirect::to('peticion/login')->with('mensaje','Debes ingresar tu cuenta para acceder');
-		   }
+	   }
+
+
+
+
+	   
+   }
+
+	  
+ 
+		  
+		  
+		  }
+		  else
+		  {
+return Redirect::to('peticion/login')->with('mensaje','Debes ingresar tu cuenta para acceder');
+		  }
 
   
-     
-     
-
-		   
-		   
-		  
-
-	
+     	
    }
 }
